+++ packages/lit-html/CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Added optional `creationScope` to `RenderOptions`, which controls the node from which the template is cloned from.
 - (Since 2.0.0-pre.6) Reintroduced the `SVGTemplateResult` type.
   ([#1623](https://github.com/Polymer/lit-html/issues/1623))